regress-svc/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
numpy==2.1.2
scikit-learn==1.5.2
joblib==1.4.2
torch==2.4.0
transformers==4.44.2
requests==2.32.3
